DB 변환 과정 개선: 경로 불일치 수정, 밝기 필터 정밀화, acos 최적화
- createHashFromDatabase.ts: data/hash → build/database 경로 수정
- reduceDatabase.ts: parseInt → parseFloat (소수점 밝기 보존)
- cell.ts: angleBetweenVectors의 Math.acos 제거 → cos 직접 비교

+++ src/scripts/database/reduceDatabase.ts
@@ -15,7 +15,7 @@ const run = ({ minBright = 10 }: { minBright: number }) => {
 };
 
 const reduce = (stars: Star[], minBright: number) => {
-  let result = stars.filter((star) => parseInt(star.V) <= minBright);
+  let result = stars.filter((star) => parseFloat(star.V) <= minBright);
   console.log(
     `밝기가 ${minBright} 보다 밝은 별들만 남깁니다. 현재 별 수: ${result.length}`,
   );
